Use fragment in Profile page
