+++ app/src/main/java/com/teamcaffeine/hotswap/navigation/ProfileFragment.java
@@ -10,15 +10,12 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
@@ -30,7 +27,6 @@ import android.widget.Toast;
 import com.facebook.share.model.ShareLinkContent;
 import com.facebook.share.widget.ShareDialog;
 import com.firebase.ui.auth.AuthUI;
-import com.google.android.gms.common.ErrorDialogFragment;
 import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
 import com.google.android.gms.common.GooglePlayServicesRepairableException;
 import com.google.android.gms.common.api.Status;
@@ -51,12 +47,10 @@ import com.teamcaffeine.hotswap.R;
 import com.teamcaffeine.hotswap.login.LoginActivity;
 import com.teamcaffeine.hotswap.login.User;
 
-
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import java.util.ArrayList;
 
 import static android.app.Activity.RESULT_CANCELED;
 import static android.app.Activity.RESULT_OK;
@@ -78,19 +72,12 @@ public class ProfileFragment extends Fragment {
     private Button btnInviteFriends;
     private TextView txtEmail;
     private TextView txtPhoneNumber;
-    private Button btnAddAddress;
-    private ListView listviewAddresses;
-    private List<String> addressElementsList;
-    private ArrayAdapter<String> addressAdapter;
+
     private Button btnAddPayment;
     private ListView listviewPayment;
     private List<String> paymentElementsList;
     private ArrayAdapter<String> paymentAdapter;
     private TextView txtPastTransactions;
-    private ListView lvAddresses;
-    private ListView lvPayment;
-//    private ListAdapter addressesAdapter;
-//    private ListAdapter paymentAdapter;
 
     public ProgressDialog mProgressDialog;
 
@@ -99,7 +86,6 @@ public class ProfileFragment extends Fragment {
     private FirebaseDatabase database;
     private DatabaseReference users;
     private String userTable = "users";
-
 
     public void showProgressDialog() {
         if (mProgressDialog == null) {
@@ -132,119 +118,8 @@ public class ProfileFragment extends Fragment {
     }
 
     @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == PLACE_AUTOCOMPLETE_REQUEST_CODE) {
-            if (resultCode == RESULT_OK) {
-                final Place place = PlaceAutocomplete.getPlace(getActivity(), data);
-
-                DatabaseReference ref = users.child(firebaseUser.getUid());
-                ref.addListenerForSingleValueEvent(new ValueEventListener() {
-                    @Override
-                    public void onDataChange(DataSnapshot dataSnapshot) {
-                        User user = dataSnapshot.getValue(User.class);
-
-                        boolean didAdd = user.addAddress(place.getAddress().toString());
-                        if (didAdd) {
-                            // Update database
-                            Map<String, Object> userUpdate = new HashMap<>();
-                            userUpdate.put(firebaseUser.getUid(), user.toMap());
-                            users.updateChildren(userUpdate);
-
-                            // Update UI
-                            addressElementsList.add(place.getAddress().toString());
-                            addressAdapter.notifyDataSetChanged();
-                        } else {
-                            Log.i(TAG, "User attempted to add a duplicate address");
-                        }
-                    }
-
-                    @Override
-                    public void onCancelled(DatabaseError databaseError) {
-                        Log.e(TAG, "Address update failed", databaseError.toException());
-                    }
-                });
-            } else if (resultCode == PlaceAutocomplete.RESULT_ERROR) {
-                Status status = PlaceAutocomplete.getStatus(getActivity(), data);
-                Log.i(TAG, status.getStatusMessage());
-                Toast.makeText(getContext(), R.string.unable_to_add_address, Toast.LENGTH_SHORT).show();
-            } else if (resultCode == RESULT_CANCELED) {
-                // The user canceled the operation.
-            }
-        }
-    }
-
-    @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-
-        btnAddAddress = view.findViewById(R.id.btnAddAddress);
-        btnAddAddress.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                try {
-                    Intent intent =
-                            new PlaceAutocomplete.IntentBuilder(PlaceAutocomplete.MODE_FULLSCREEN)
-                                    .build(getActivity());
-                    startActivityForResult(intent, PLACE_AUTOCOMPLETE_REQUEST_CODE);
-                } catch (GooglePlayServicesRepairableException | GooglePlayServicesNotAvailableException e) {
-                    Log.e(TAG, "Google Places Error", e);
-                }
-            }
-        });
-
-        listviewAddresses = view.findViewById(R.id.listviewAddresses);
-        listviewAddresses.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, final int position, long id) {
-                AlertDialog myQuittingDialogBox = new AlertDialog.Builder(getContext())
-                        //set message, title, and icon
-                        .setTitle(R.string.delete)
-                        .setMessage(R.string.delete_address_question)
-                        .setPositiveButton(R.string.delete, new DialogInterface.OnClickListener() {
-
-                            public void onClick(DialogInterface dialog, int whichButton) {
-                                DatabaseReference ref = users.child(firebaseUser.getUid());
-                                ref.addListenerForSingleValueEvent(new ValueEventListener() {
-                                    @Override
-                                    public void onDataChange(DataSnapshot dataSnapshot) {
-                                        User user = dataSnapshot.getValue(User.class);
-
-                                        boolean didRemove = user.removeAddress(listviewAddresses.getItemAtPosition(position).toString());
-                                        if (didRemove) {
-                                            // Update database
-                                            Map<String, Object> userUpdate = new HashMap<>();
-                                            userUpdate.put(firebaseUser.getUid(), user.toMap());
-                                            users.updateChildren(userUpdate);
-
-                                            // Update UI
-                                            addressElementsList.remove(position);
-                                            addressAdapter.notifyDataSetChanged();
-                                        } else {
-                                            Log.i(TAG, "User attempted to delete a nonexistent address");
-                                        }
-                                    }
-
-                                    @Override
-                                    public void onCancelled(DatabaseError databaseError) {
-                                        Log.e(TAG, "Address update failed", databaseError.toException());
-                                    }
-                                });
-                                dialog.dismiss();
-                            }
-                        })
-                        .setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int which) {
-                                dialog.dismiss();
-                            }
-                        })
-                        .create();
-                myQuittingDialogBox.show();
-            }
-        });
-
-
-
 
         btnAddPayment = view.findViewById(R.id.btnAddPayment);
         btnAddPayment.setOnClickListener(new View.OnClickListener() {
@@ -336,12 +211,6 @@ public class ProfileFragment extends Fragment {
 
                 txtEmail.setText(user.getEmail());
                 txtPhoneNumber.setText(user.getPhoneNumber());
-
-                addressElementsList = user.getAddresses();
-                addressAdapter = new ArrayAdapter<String>
-                        (getContext(), android.R.layout.simple_list_item_1, addressElementsList);
-                listviewAddresses.setAdapter(addressAdapter);
-                addressAdapter.notifyDataSetChanged();
 
                 paymentElementsList = user.getPayments();
                 paymentAdapter = new ArrayAdapter<String>
