+++ app/src/main/res/layout/activity_profile.xml
@@ -93,7 +93,7 @@
         </LinearLayout>
     </LinearLayout>
 
-    <LinearLayout
+    <FrameLayout
         android:id="@+id/layout_Adresses"
         android:layout_width="match_parent"
         android:layout_height="50dp"
@@ -101,43 +101,12 @@
         android:orientation="horizontal"
         android:weightSum="5">
 
-        <LinearLayout
-            android:layout_width="0dp"
+        <fragment
+            android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:orientation="vertical">
+            class="com.teamcaffeine.hotswap.navigation.AddressesFragment" />
 
-            <TextView
-                android:id="@+id/txtAddressesLabel"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/addresses_label" />
-
-            <Button
-                android:id="@+id/btnAddAddress"
-                style="@android:style/Widget.Material.Light.Button.Small"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/plus"
-                android:textColor="@color/blue" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/layout_AddressesAndAdd"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="4"
-            android:orientation="vertical">
-
-            <ListView
-                android:id="@+id/listviewAddresses"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="1"></ListView>
-
-        </LinearLayout>
-
-    </LinearLayout>
+    </FrameLayout>
 
     <LinearLayout
         android:id="@+id/layout_Payment"
